requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.9
django-rosetta==0.9.9
idna==3.4
Pillow==9.5.0
polib==1.2.0
requests==2.29.0
rosetta==0.3
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
